backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.2.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
drf-spectacular==0.27.1
dj-database-url==2.1.0
